move get_fee to ledger
