+++ xrpl/account/__init__.py
@@ -1,10 +1,5 @@
 """Public interface for XRPL account sugar methods."""
 
-from xrpl.account.main import (
-    get_account_info,
-    get_balance,
-    get_fee,
-    get_next_valid_seq_number,
-)
+from xrpl.account.main import get_account_info, get_balance, get_next_valid_seq_number
 
-__all__ = ["get_fee", "get_next_valid_seq_number", "get_balance", "get_account_info"]
+__all__ = ["get_next_valid_seq_number", "get_balance", "get_account_info"]
