+++ xrpl/ledger/__init__.py
@@ -1,4 +1,4 @@
 """Top-level exports for the ledger package."""
-from xrpl.ledger.main import get_latest_validated_ledger_sequence
+from xrpl.ledger.main import get_fee, get_latest_validated_ledger_sequence
 
-__all__ = ["get_latest_validated_ledger_sequence"]
+__all__ = ["get_latest_validated_ledger_sequence", "get_fee"]
