+++ xrpl/ledger/main.py
@@ -3,7 +3,7 @@
 from typing import Any, Dict, cast
 
 from xrpl.clients import Client
-from xrpl.models.requests import Ledger
+from xrpl.models.requests import Fee, Ledger
 
 
 def get_latest_validated_ledger_sequence(client: Client) -> int:
@@ -19,3 +19,18 @@ def get_latest_validated_ledger_sequence(client: Client) -> int:
     response = client.request(Ledger(ledger_index="validated"))
     result = cast(Dict[str, Any], response.result)
     return cast(int, result["ledger_index"])
+
+
+def get_fee(client: Client) -> str:
+    """
+    Query the ledger for the current minimum transaction fee.
+
+    Args:
+        client: the network client used to make network calls.
+
+    Returns:
+        The minimum fee for transactions.
+    """
+    response = client.request(Fee())
+    result = cast(Dict[str, Any], response.result)
+    return cast(str, result["drops"]["minimum_fee"])
